backend/requirements.txt
Flask
mysql-connector-python
pandas
matplotlib
seaborn
plotly
numpy
requests
gunicorn